requirements.txt
streamlit
tensorflow
keras
matplotlib
pandas
numpy
opencv-python
pillow